# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이화여대 창업 교육 LMS API",
        "title": "EarnLearning LMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "earnlearning.com",
    "basePath": "/api",
    "paths": {
        "/admin/classrooms/{id}/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "관리자용: 클래스룸 멤버 활동 대시보드",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "클래스룸 대시보드",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "클래스룸 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/classrooms": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내가 속한 클래스룸 목록 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classroom"
                ],
                "summary": "내 클래스룸 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "새 클래스룸 생성 (관리자)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classroom"
                ],
                "summary": "클래스룸 생성",
                "parameters": [
                    {
                        "description": "클래스룸 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.CreateClassroomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/classrooms/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "초대 코드로 클래스룸 참여",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classroom"
                ],
                "summary": "클래스룸 참여",
                "parameters": [
                    {
                        "description": "참여 코드",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.JoinClassroomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/classrooms/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "클래스룸 정보, 채널, 멤버 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classroom"
                ],
                "summary": "클래스룸 상세 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "클래스룸 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_interfaces_http_handler.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_INPUT"
                },
                "message": {
                    "type": "string",
                    "example": "잘못된 입력입니다"
                }
            }
        },
        "internal_interfaces_http_handler.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_interfaces_http_handler.APIError"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "internal_interfaces_http_handler.CreateClassroomRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "코딩입문 1반"
                }
            }
        },
...
        "internal_interfaces_http_handler.JoinClassroomRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ABC123"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}